Split agent-page.tsx into per-tab files
- activity-panel.tsx: ActivityPanel + helpers
- knowledge-panel.tsx: KnowledgePanel (formerly SkillsPanel)
- permissions-panel.tsx: PermissionsPanel (formerly ToolsPanel)
- Delete agent-page.tsx barrel file
